Post review

+++ include/defines/CameraDefines.hpp
@@ -1,7 +1,7 @@
 #ifndef CAMERADEFINES_HPP
 #define CAMERADEFINES_HPP
 
-#define FOV 90.0f                                   // Field of view
+#define FOV static_cast<float>(90)                  // Field of view (in degrees)
 #define NEAR_PLANE_Z 0.1f                           // Near clipping plane
 #define FAR_PLANE_Z 100.0f                          // Far clipping plane
 #define Z_RANGE (NEAR_PLANE_Z - FAR_PLANE_Z)        // Range of the z-axis
